chore(keyboard preview): refactor for settings
